requirements.txt
python-telegram-bot[job-queue]>=22.0
rapidfuzz>=3.12
confusable_homoglyphs>=3.3
aiosqlite>=0.21
pyyaml>=6.0
python-dotenv>=1.1

# Testing
pytest>=8.3
pytest-asyncio>=0.25
